requirements.txt
asgiref==3.7.2
boto3==1.35.16
channels==3.0.5
channels-redis==4.2.0
Django==4.2.11
django-filter==24.3
django-enumfields==2.1.1
django-model-utils==4.5.1
django-storages==1.14.4
django-passwords==0.3.12
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.7.0
migrate==0.3.8
phonenumbers==8.13.31
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
pytz==2024.1
pillow==10.4.0
sqlparse==0.4.4
typing_extensions==4.10.0
weasyprint==63.1
daphne==3.0
requests==2.32.3
openpyxl==3.1.5
sentry-sdk==2.26.1
whitenoise==6.2.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
